+++ src/factory/factory.h
@@ -24,12 +24,12 @@ extern zend_class_entry *gene_factory_ce;
 GENE_MINIT_FUNCTION(factory);
 
 
-zend_bool gene_factory(char *className, int tmp_len, zval *params, zval *classObject);
+zend_bool gene_factory(char *className, size_t tmp_len, zval *params, zval *classObject);
 void gene_factory_call(zval *object, char *action, zval *param, zval *retval);
 void gene_factory_call_1(zval *object, char *action, zval *param, zval *retval);
 void gene_factory_call_2(char *method, zval *key, zval *retval);
 void gene_factory_function_call(char *action, zval *param_key, zval *param_arr, zval *retval);
 void gene_factory_function_call_1(zval *function_name, zval *param_key, zval *param_arr, zval *retval);
-zend_bool gene_factory_load_class(char *className, int tmp_len, zval *classObject);
+zend_bool gene_factory_load_class(char *className, size_t tmp_len, zval *classObject);
 
 #endif
